Move notificationfield into its own file
